requirements.txt
pandas
numpy
sklearn
matplotlib
seaborn
fastapi
gunicorn
mlflow
jupyter
pandas-profiling
boto3
psycopg2-binary
python-decouple